Final Commit(AGAIN!!!)
Discovered issues with radio buttons, when selection of radio button changed, so did the first list. When radio was changed multiple times it was reading items to the list because the radio was tied to a method that added objects to the list. Rectified this by adding a separate method that populated the list to the radio selection change method. Also, ensured that when an item was removed from a list, and the radio button changed, the item did not appear in any of the filtered lists.